--- src/main/java/kr/hhplus/be/server/order/request/OrderProductRequest.java
@@ -1,22 +1,0 @@
-package kr.hhplus.be.server.order.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class OrderProductRequest {
-
-    private long productId;
-
-    private long optionId;
-
-    private long productPrice;
-
-    private long orderQty;
-
-    public static OrderProductRequest of(long productId, long optionId,long productPrice, long orderQty) {
-        return new OrderProductRequest(productId, optionId, productPrice, orderQty);
-    }
-
-}
