--- src/main/java/kr/hhplus/be/server/user/response/UserResponse.java
@@ -1,17 +1,0 @@
-package kr.hhplus.be.server.user.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserResponse {
-
-    private long userId;
-
-    private long point;
-
-    public static UserResponse of(long userid, long point){
-        return new UserResponse(userid, point);
-    }
-}
