+++ src/main/java/kr/hhplus/be/server/web/order/request/OrderRequest.java
@@ -1,0 +1,23 @@
+package kr.hhplus.be.server.web.order.request;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import java.util.List;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class OrderRequest {
+
+    private long userId;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private long couponId;
+
+    List<OrderProductRequest> items;
+
+    public static OrderRequest of(long userId ,long couponId, List<OrderProductRequest> items) {
+        return new OrderRequest(userId, couponId, items);
+    }
+
+}
