+++ src/main/java/kr/hhplus/be/server/web/user/SwaggerUserResponseResult.java
@@ -1,0 +1,23 @@
+package kr.hhplus.be.server.web.user;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SwaggerUserResponseResult {
+
+
+    @Schema(description = "code", example = "200")
+    private int code;
+
+    @Schema(description = "message", example = "success")
+    private String message;
+
+    @Schema(
+        description = "잔액 데이터",
+        example = "{\"user_id\":1,\"amount\":10000}"
+    )
+    private UserResponse data;
+}
