+++ src/main/java/kr/hhplus/be/server/web/user/UserResponse.java
@@ -1,0 +1,17 @@
+package kr.hhplus.be.server.web.user;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserResponse {
+
+    private long userId;
+
+    private long point;
+
+    public static UserResponse of(long userid, long point){
+        return new UserResponse(userid, point);
+    }
+}
